Jumbo Blob support (#1113)
* Support Jumbo Blobs

* Jumbo blobs: Improve chunking logic

* Improve chunking logic

* Review Ze's comments

* Add test to verify block size parameters

* Improve error messaging and STE data schema version

+++ ste/xfer-remoteToLocal-file.go
@@ -53,7 +53,7 @@ func remoteToLocal_file(jptm IJobPartTransferMgr, p pipeline.Pipeline, pacer pac
 
 	// step 2: get the source, destination info for the transfer.
 	fileSize := int64(info.SourceSize)
-	downloadChunkSize := int64(info.BlockSize)
+	downloadChunkSize := info.BlockSize
 
 	// step 3: Perform initial checks
 	// If the transfer was cancelled, then report transfer as done
